% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{uv_add}
\alias{uv_add}
\title{Add serie}
\usage{
uv_add(p, serie)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{serie}{values to plot.}
}
\description{
Add serie
}
\examples{
mtcars \%>\%
  uv_bar(mpg) \%>\%
  uv_add(qsec) \%>\%
  uv_add(drat)

}

