% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_areaopt}
\alias{uv_areaopt}
\title{Customise chart lines}
\usage{
uv_areaopt(p, interpolation, opacity, offset)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{interpolation}{line interpolation. Defaults to \code{linear} see details for valid values.}

\item{opacity}{oOpacity in the case of simple Area chart, defaults to \code{0.2}.}

\item{offset}{offset, defaults to \code{zero}. See details for valid values.}
}
\description{
Customise the properties specific to line chart.
}
\details{
Valid values for \code{interpolation} argument:
\itemize{
  \item{\code{linear}}
  \item{\code{basis}}
  \item{\code{cardinal}}
  \item{\code{monotone}}
}

Valid values for \code{offset} argument:
\itemize{
  \item{\code{zero}}
  \item{\code{wiggle}}
  \item{\code{silhouette}}
  \item{\code{expand}}
}
}
\examples{
mtcars \%>\%
  uv_area(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_areaopt(opacity = 0.9)

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

