% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\name{uv_axis}
\alias{uv_axis}
\title{Customise chart axis}
\usage{
uv_axis(p, opacity, ticks, subticks, padding, minor, strokecolor, fontfamily,
  fontsize, fontweight, showticks, showsubticks, showtext)
}
\description{
Customise the chart axis
}

