% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_baropt}
\alias{uv_baropt}
\title{Customise chart bar chart}
\usage{
uv_baropt(p, strokecolor, fontsize, fontweight, fontfamily, textcolor)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{strokecolor}{any Color Code or Color String or \code{none}. defaults to \code{#000000}.}

\item{fontsize}{font size. Defaults to \code{11}.}

\item{fontweight}{font weight, defaults to \code{bold}.}

\item{fontfamily}{any font, defaults to \code{Arial}.}

\item{textcolor}{text color used for the numbers, defaults to \code{#000000}.}
}
\description{
Customise the chart bars.
}
\examples{
mtcars \%>\%
  uv_stackbar(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_add(drat) \%>\%
  uv_baropt(fontsize = 7)

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

