% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_caption}
\alias{uv_caption}
\alias{uv_subcaption}
\title{Customise chart caption}
\usage{
uv_caption(p, fontfamily, fontsize, fontweight, textdecoration, strokecolor,
  cursor)

uv_subcaption(p, fontfamily, fontsize, fontweight, textdecoration, strokecolor,
  cursor)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{fontfamily}{any font, defaults to \code{Arial}.}

\item{fontsize}{font size. Defaults to \code{11}.}

\item{fontweight}{font weight, defaults to \code{bold}.}

\item{textdecoration}{text decoration applied on the caption, defaults to \code{none}.}

\item{strokecolor}{any Color Code or Color String or \code{none}. defaults to \code{#000000}.}

\item{cursor}{precision on the labels in case of decimal values. Defaults to \code{pointer}.}
}
\description{
Customise the chart caption.

Customise the chart subcaption.
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

