% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvcharts.R
\name{uv_charts}
\alias{uv_area}
\alias{uv_bar}
\alias{uv_charts}
\alias{uv_donut}
\alias{uv_line}
\alias{uv_percentarea}
\alias{uv_percentbar}
\alias{uv_pie}
\alias{uv_polar}
\alias{uv_stackarea}
\alias{uv_stackbar}
\alias{uv_stepupbar}
\alias{uv_waterfall}
\title{Setup a uvchart}
\usage{
uv_charts(data, x, type = "Line", width = NULL, height = NULL,
  elementId = NULL)

uv_bar(data, x, width = NULL, height = NULL, elementId = NULL)

uv_line(data, x, width = NULL, height = NULL, elementId = NULL)

uv_area(data, x, width = NULL, height = NULL, elementId = NULL)

uv_stackbar(data, x, width = NULL, height = NULL, elementId = NULL)

uv_stackarea(data, x, width = NULL, height = NULL, elementId = NULL)

uv_pie(data, x, width = NULL, height = NULL, elementId = NULL)

uv_donut(data, x, width = NULL, height = NULL, elementId = NULL)

uv_percentbar(data, x, width = NULL, height = NULL, elementId = NULL)

uv_percentarea(data, x, width = NULL, height = NULL, elementId = NULL)

uv_stepupbar(data, x, width = NULL, height = NULL, elementId = NULL)

uv_polar(data, x, width = NULL, height = NULL, elementId = NULL)

uv_waterfall(data, x, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{data.frame of data to plot.}

\item{x}{\code{x} variable.}

\item{type}{type of chart to plot, see details.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{id of div element containing chart.}
}
\description{
Initiate a uvchart
}

