% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_dimensions}
\alias{uv_dimensions}
\title{Customise chart dimensions}
\usage{
uv_dimensions(p, width, height)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{width, height}{any CSS size.}
}
\description{
Customise the chart dimensions.
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

