% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_frame}
\alias{uv_frame}
\title{Customise chart frame}
\usage{
uv_frame(p, bgcolor)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{bgcolor}{any Color Code or Color String or \code{none}.}
}
\description{
Customise the chart frame.
}
\examples{
mtcars \%>\%
  uv_charts(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_frame(bgcolor = "black")

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

