% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_graph}
\alias{uv_graph}
\title{Customise chart caption}
\usage{
uv_graph(p, palette, bgcolor, orientation, custom.palette, opacity,
  responsive = TRUE)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{palette}{palette name, see details for valid values. Defaults to \code{Brink}.}

\item{bgcolor}{background color used for the entire chart include the margin spaces.}

\item{orientation}{orientation of the chart for a different layout based on requirements. Can take
\code{Horizontal} or \code{Vertical}, defaults to the former.}

\item{custom.palette}{override the palette with your own color palette by providing a non empty array
of color codes. They are used in order for corresponding category in the categories array of the graph
definition. Pass colors as \code{list}.}

\item{opacity}{opacity of the entire chart to make it blend into any background of your choice. Defaults to \code{1}.}

\item{responsive}{\code{TRUE} for responsive sizing (default).}
}
\description{
Customise the chart caption.
}
\details{
Valid values for \code{palette} argument.

\itemize{
  \item{\code{Plain}}
  \item{\code{Simple}}
  \item{\code{Olive}}
  \item{\code{Candid}}
  \item{\code{Sulphide}}
  \item{\code{Soft}}
  \item{\code{Lint}}
}
}
\examples{
mtcars \%>\%
  uv_charts(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_graph(palette = "Plain")

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

