% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_label}
\alias{uv_label}
\title{Customise chart labels}
\usage{
uv_label(p, fontfamily, fontsize, fontweight, strokecolor, showlabel, precision,
  prefix, suffix)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{fontfamily}{any font, defaults to \code{Arial}.}

\item{fontsize}{font size. Defaults to \code{11}.}

\item{fontweight}{font weight, defaults to \code{bold}.}

\item{strokecolor}{any Color Code or Color String or \code{none}. defaults to \code{#000000}.}

\item{showlabel}{show labels flag, \code{TRUE} or \code{FALSE}, defaults to the former.}

\item{precision}{precision on the labels in case of decimal values. Defaults to \code{2}.}

\item{prefix}{any prefix which needs to be added to the label.}

\item{suffix}{any suffix which needs to be added to the label.}
}
\description{
Customise the chart labels.
}
\examples{
mtcars \%>\%
  uv_charts(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_label(fontsize = 6, strokecolor = "red")

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

