% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_lineopt}
\alias{uv_lineopt}
\title{Customise chart lines}
\usage{
uv_lineopt(p, interpolation)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{interpolation}{line interpolation. Defaults to \code{linear} see details for valid values.}
}
\description{
Customise the properties affecting Bar, Stacked Bar and Step Up Bar Chart.
}
\details{
Valid values for \code{interpolation} argument:
\itemize{
  \item{\code{linear}}
  \item{\code{basis}}
  \item{\code{cardinal}}
  \item{\code{monotone}}
}
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

