% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_margin}
\alias{uv_margin}
\title{Customise chart margin}
\usage{
uv_margin(p, top, bottom, left, right)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{top, bottom, left, right}{any CSS size.}
}
\description{
Customise the chart margin.
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

