% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_meta}
\alias{uv_meta}
\title{Customise chart meta properties}
\usage{
uv_meta(p, caption, subcaption, hlabel, vlabel, hsulabel, vsulabel,
  isDownloadable, downloadLabel)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{caption}{caption for the graph.}

\item{subcaption}{subcaption of the graph.}

\item{hlabel}{label for the horizontal axis.}

\item{vlabel}{label for the vertical axis.}

\item{hsulabel}{sub-label for the horizontal axis.}

\item{vsulabel}{sub-label for the vertical axis.}

\item{isDownloadable}{whether or not to display the download link for the chart, defaults to \code{FALSE}.}

\item{downloadLabel}{text to be displayed in the download link, defaults to \code{Download}.}
}
\description{
Customise the chart meta properties.
}
\examples{
mtcars \%>\%
  uv_charts(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_meta("Title", "Subtitle", "x axis", "y axis")

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

