% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_pieopt}
\alias{uv_donutopt}
\alias{uv_pieopt}
\title{Customise chart lines}
\usage{
uv_pieopt(p, fontfamily, fontsize, fontweight, fontvariant, fontfill,
  strokecolor, strokewidth)

uv_donutopt(p, fontfamily, fontsize, fontweight, fontvariant, fontfill,
  strokecolor, strokewidth, factor)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{fontfamily}{any font, defaults to \code{Arial}.}

\item{fontsize}{font size. Defaults to \code{11}.}

\item{fontweight}{font weight, defaults to \code{bold}.}

\item{fontvariant}{font variant used on the text, defaults to \code{small-caps}, can also take \code{normal}.}

\item{fontfill}{text color on the measure, defaults to \code{#000000}.}

\item{strokecolor}{any Color Code or Color String or \code{none}. defaults to \code{#000000}.}

\item{strokewidth}{stroke width on each arc. Defaults to \code{1}.}

\item{factor}{radius of the cropped part in terms of the bigger pie. Defaults to \code{0.4}.}
}
\description{
Customise the properties specific to Pie Chart.

Customise the properties specific to Donut Chart.
}
\examples{
mtcars \%>\%
  uv_pie(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_pieopt(fontfill = "white")

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}

\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

