% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_scale}
\alias{uv_scale}
\title{Customise chart scale}
\usage{
uv_scale(p, type, ordinality)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{type}{scale used to represent values on the chart, see details for valid values.}

\item{ordinality}{defines the gap between 2 consecutive labels, defaults to \code{0.2}.}
}
\description{
Customise the chart scale.
}
\details{
Valid values for \code{type} argument, defaults to \code{linear}:

\itemize{
  \item{\code{linear}}
  \item{\code{log}}
  \item{\code{pow}}
  \item{\code{sqrt}}
}
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

