% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_tooltip}
\alias{uv_tooltip}
\title{Customise chart tooltip}
\usage{
uv_tooltip(p, show, format)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{show}{flag to enable or disable tooltip functionality, defaults to \code{TRUE}.}

\item{format}{defines custom format for tooltip, character string. defaults to \code{\%c [\%l] : \%v}}
}
\description{
Customise the chart tooltip.
}
\examples{
mtcars \%>\%
  uv_charts(qsec) \%>\%
  uv_add(mpg) \%>\%
  uv_tooltip(format = "\%v")

}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

