% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_axis}
\alias{uv_axis}
\title{Customise chart axis}
\usage{
uv_axis(p, opacity, ticks, subticks, padding, minor, strokecolor, fontfamily,
  fontsize, fontweight, showticks, showsubticks, showtext)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{opacity}{opacity of the axis lines drawn for reference, defaults to \code{1}.}

\item{ticks}{number of major ticks on the measure axis.}

\item{subticks}{number of minor ticks on the measure axis.}

\item{padding}{any CSS size.}

\item{minor}{length of the minor ticks towards the label.}

\item{strokecolor}{color of axis lines, any Color Code or Color String or \code{none}. defaults to \code{#000000}.}

\item{fontfamily}{any font, defaults to \code{Arial}.}

\item{fontsize}{font size. Defaults to \code{11}.}

\item{fontweight}{font weight, defaults to \code{bold}.}

\item{showticks}{show ticks flag, \code{TRUE} or \code{FALSE}, defaults to the former.}

\item{showsubticks}{show subticks flag, \code{TRUE} or \code{FALSE}, defaults to the former.}

\item{showtext}{show text flag, \code{TRUE} or \code{FALSE}, defaults to the former.}
}
\description{
Customise the chart axis.
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

