% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_config}
\alias{uv_config}
\title{Configure the chart}
\usage{
uv_config(p, ...)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{...}{any argument to pass to \href{http://imaginea.github.io/uvCharts/documentation.html}{config}.}
}
\description{
Configure the chart.
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

