% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{uv_legend}
\alias{uv_legend}
\title{Configure the chart}
\usage{
uv_legend(p, showlegends, position)
}
\arguments{
\item{p}{a \code{uvcharts} object.}

\item{showlegends}{boolean.}

\item{position}{legend position, i.e.:\code{right}.}
}
\description{
Configure the chart.
}
\seealso{
\href{http://imaginea.github.io/uvCharts/documentation.html}{Official docs}
}

